Add a test for _WKInputDelegate.willSubmitFormValues https://bugs.webkit.org/show_bug.cgi?id=179086
Reviewed by Darin Adler.

Hidden input types are not included in this submission because FormSubmission::create ignores them
because input.isTextField() returns false.  This is existing behavior that wasn't shown by a test.
HTTPBodies are also not sent to WKURLSchemeHandlers.  This is existing behavior that wasn't shown by a test.

* TestWebKitAPI/TestWebKitAPI.xcodeproj/project.pbxproj:
* TestWebKitAPI/Tests/WebKitCocoa/_WKInputDelegate.mm: Added.
(-[FormSubmissionDelegate _webView:willSubmitFormValues:userObject:submissionHandler:]):
(TEST):



Identifier: 195232@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@224281 268f45cc-cd09-0410-ab3c-d52691b4dbfc