Don't eliminate whitespace renderer if the previous sibling is a text renderer https://bugs.webkit.org/show_bug.cgi?id=179620
Reviewed by Zalan Bujtas.

Source/WebCore:

Currently whitespace elimination code doesn't consider runs of text renderers. We should always make whitespace
renderer if the previous renderer is a text renderer. The behavior should be the same as if those were a single
renderer with merged text. This situation can happen easily with display:contents.

This fixes the remaining flexbox failures in display:contents tests.

* style/RenderTreeUpdater.cpp:
(WebCore::RenderTreeUpdater::textRendererIsNeeded):

We need a renderer if the previous rendere is RenderText.

LayoutTests:

* TestExpectations:


Identifier: 195660@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@224773 268f45cc-cd09-0410-ab3c-d52691b4dbfc