Crash in WebCore::SWServerJobQueue::scriptFetchFinished(WebCore::SWServer::Connection&, WebCore::ServiceWorkerFetchResult const&) + 77 https://bugs.webkit.org/show_bug.cgi?id=179394 <rdar://problem/35478190>
Reviewed by Youenn Fablet.

In SWServerJobQueue::scriptFetchFinished(), make sure the job queue is still processing the
job that triggered the fetch. If it is not, ignore the callback.

This could happen when calling SWServer::clear() while a script fetch was pending.

No new tests, already covered by existing tests that sometimes crash on the bots.

* workers/service/server/SWServerJobQueue.cpp:
(WebCore::SWServerJobQueue::isCurrentlyProcessingJob const):
(WebCore::SWServerJobQueue::scriptFetchFinished):
* workers/service/server/SWServerJobQueue.h:

Identifier: 195819@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@224943 268f45cc-cd09-0410-ab3c-d52691b4dbfc