Dispatching an event on a ServiceWorkerRegistration may fail or crash due to GC https://bugs.webkit.org/show_bug.cgi?id=179751
Reviewed by Geoffrey Garen.

Source/WebCore:

Dispatching an event on a ServiceWorkerRegistration may fail or crash due to GC. We
need to make sure that a ServiceWorkerRegistration's wrapper stays alive as long as
we may dispatch events on it.

Keep the wrapper alive by making ServiceWorker an ActiveDOMObject and making sure the
implementation object keeps a PendingActivity alive while it may dispatch JS events.
The only event dispatched on ServiceWorkerRegistration objects is the "updatefound"
one. We may dispatch updatefound events on a ServiceWorkerRegistration until the
registration gets cleared on server side, in which case the ServiceWorkerRegistration
object ends up with no service worker. We therefore take a PendingActivity when the
ServiceWorkerRegistration is constructed and we release it once the registration has
no more associated ServiceWorker or the ActiveDOMObject is stopped (to avoid leaks on
navigation).

Test: http/tests/workers/service/service-worker-registration-gc-event.html

* workers/service/ServiceWorkerContainer.cpp:
(WebCore::ServiceWorkerContainer::scheduleTaskToFireUpdateFoundEvent):
* workers/service/ServiceWorkerRegistration.cpp:
(WebCore::ServiceWorkerRegistration::getOrCreate):
(WebCore::ServiceWorkerRegistration::updateStateFromServer):
(WebCore::ServiceWorkerRegistration::scheduleTaskToFireUpdateFoundEvent):
(WebCore::ServiceWorkerRegistration::stop):
(WebCore::ServiceWorkerRegistration::updatePendingActivityForEventDispatch):
* workers/service/ServiceWorkerRegistration.h:
* workers/service/ServiceWorkerRegistration.idl:

LayoutTests:

Add layout test coverage.

* http/tests/workers/service/service-worker-registration-gc-event-expected.txt: Added.
* http/tests/workers/service/service-worker-registration-gc-event.html: Added.

Identifier: 195820@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@224944 268f45cc-cd09-0410-ab3c-d52691b4dbfc